feat: add new form api endpoint for getting form stuff
