+++ core/app/components/DataTable/PubsDataTable/FormPubsDataTableClient.tsx
@@ -1,0 +1,72 @@
+"use client";
+
+import { useState } from "react";
+
+import type { NonGenericProcessedPub } from "contracts";
+import type { FormsId, PubsId, PubTypes } from "db/public";
+
+import { client } from "~/lib/api";
+import { type GetManyParams } from "~/lib/server";
+import { useFormData } from "../../forms/FormDataProvider";
+import { useCommunity } from "../../providers/CommunityProvider";
+import { PubsDataTableClient, PubsDataTableClientBase } from "./PubsDataTableClient";
+
+type FormPubsDataTableClientProps = {
+	/* The slug of the field on the form that's being used to render the table */
+	fieldSlug: string;
+	selectedPubs?: NonGenericProcessedPub[];
+	onSelectedPubsChange?: (pubs: NonGenericProcessedPub[]) => void;
+	disabledRows?: PubsId[];
+	pubTypes?: Pick<PubTypes, "id" | "name">[];
+	currentPubId?: PubsId;
+};
+
+/**
+ * Form-context aware version of PubsDataTableClient that uses the secure form endpoint.
+ * This component can only be used within a form context and requires a valid form token.
+ */
+export const FormPubsDataTableClient = (props: FormPubsDataTableClientProps) => {
+	const [filterParams, setFilterParams] = useState<Required<GetManyParams>>({
+		limit: 10,
+		offset: 0,
+		orderBy: "updatedAt",
+		orderDirection: "desc",
+	});
+
+	const { form, token } = useFormData();
+
+	const community = useCommunity();
+
+	const { data, isLoading } = client.site.forms.getPubsForFormField.useQuery({
+		queryKey: ["getPubs", filterParams],
+		queryData: {
+			query: {
+				...filterParams,
+				pubTypeId: props.pubTypes ? props.pubTypes.map((p) => p.id) : undefined,
+				withPubType: true,
+				withRelatedPubs: false,
+				withStage: true,
+				withValues: false,
+			},
+			params: {
+				communitySlug: community.slug,
+				formSlug: form.slug,
+				fieldSlug: props.fieldSlug,
+				pubId: props.currentPubId,
+			},
+			// headers: {
+			// 	"x-form-token": token,
+			// },
+		},
+	});
+
+	return (
+		<PubsDataTableClientBase
+			{...props}
+			data={data}
+			isLoading={isLoading}
+			filterParams={filterParams}
+			setFilterParams={setFilterParams}
+		/>
+	);
+};
