+++ core/app/components/FormSwitcher/FormSwitcher.tsx
@@ -2,6 +2,7 @@
 
 import { useEffect, useState } from "react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { parseAsString, useQueryState } from "nuqs";
 
 import {
 	Select,
@@ -30,18 +31,24 @@ export const FormSwitcher = ({
 	className?: string;
 	children?: React.ReactNode;
 }) => {
-	const router = useRouter();
-	const pathname = usePathname();
-	const params = useSearchParams();
+	// const router = useRouter();
+	// const pathname = usePathname();
+	// const params = useSearchParams();
 
-	const [selectedFormSlug, setSelectedFormSlug] = useState(defaultFormSlug ?? forms[0].slug);
-	useEffect(() => {
-		if (!params.get(formSwitcherUrlParam)) {
-			const newParams = new URLSearchParams(params);
-			newParams.set(formSwitcherUrlParam, selectedFormSlug);
-			router.replace(`${pathname}?${newParams.toString()}`, { scroll: false });
-		}
-	}, []);
+	// const [selectedFormSlug, setSelectedFormSlug] = useState(defaultFormSlug ?? forms[0].slug);
+	// useEffect(() => {
+	// 	if (!params.get(formSwitcherUrlParam)) {
+	// 		const newParams = new URLSearchParams(params);
+	// 		newParams.set(formSwitcherUrlParam, selectedFormSlug);
+	// 		router.replace(`${pathname}?${newParams.toString()}`, { scroll: false });
+	// 	}
+	// }, []);
+	const [selectedFormSlug, setSelectedFormSlug] = useQueryState(
+		formSwitcherUrlParam,
+		parseAsString.withDefault(defaultFormSlug ?? forms[0].slug).withOptions({
+			shallow: false,
+		})
+	);
 
 	if (!forms.length) {
 		return null;
